Extract clinker-core-types crate for shared leaf vocabulary
Introduce a new workspace crate, clinker-core-types, that holds the
value-level vocabulary every higher pipeline layer passes around but that
carries no executor, config, or schema coupling: source spans
(FileId, Span), the structured compile-time diagnostic types (Diagnostic,
Severity, LabeledSpan, DiagnosticPayload), the name-keyed DAG graph used
for cycle detection (NameGraph), and the dead-letter-queue category enum
plus its stage-label helpers (DlqErrorCategory, stage_aggregate,
stage_time_window). The crate sits below clinker-core and depends only on
miette, serde-saphyr, and petgraph, so every higher crate can use these
types without a dependency cycle.

The orchestration-coupled members of the original modules stay in
clinker-core, since they still reference its config, schema, executor,
pipeline, and path-security types:

- PipelineError keeps its From conversions over ConfigError, SchemaError,
  SpillError, and the memory BudgetCategory.
- The DLQ CSV writers (write_dlq, partition_dlq_entries) keep their
  dependency on DlqConfig and DlqEntry.
- SourceDb keeps its load() path through the ValidatedPath security
  token, importing the leaf FileId/Span value types from the new crate.

Every clinker-core reference to the relocated symbols is repointed at
clinker_core_types directly, and the clinker-channel and clinker
consumers that used the diagnostic and span value types are repointed and
given a direct dependency on the new crate. No re-export facade or
parallel import path is left behind.
