+++ crates/clinker-channel/Cargo.toml
@@ -7,6 +7,7 @@ description = "Channel binding, validation, and overlay merge for Clinker"
 
 [dependencies]
 clinker-core = { workspace = true }
+clinker-core-types = { workspace = true }
 clinker-record = { workspace = true }
 serde-saphyr = { workspace = true }
 serde = { workspace = true }
