+++ crates/clinker-core/tests/combine_build_direct_rollback.rs
@@ -27,9 +27,9 @@ use std::path::PathBuf;
 
 use clinker_bench_support::io::SharedBuffer;
 use clinker_core::config::{CompileContext, parse_config};
-use clinker_core::dlq::DlqErrorCategory;
 use clinker_core::executor::{DlqEntry, PipelineExecutor, PipelineRunParams, SourceReaders};
 use clinker_core::source::multi_file::FileSlot;
+use clinker_core_types::dlq::DlqErrorCategory;
 
 fn slot(name: &str, csv: &str) -> FileSlot {
     FileSlot::new(
