+++ crates/clinker-core/tests/time_window_aggregate.rs
@@ -28,8 +28,8 @@ use std::io::{self, Cursor, Write};
 use std::sync::{Arc, Mutex};
 
 use clinker_core::config::{CompileContext, PipelineConfig, parse_config};
-use clinker_core::dlq::DlqErrorCategory;
 use clinker_core::executor::{ExecutionReport, PipelineExecutor, PipelineRunParams};
+use clinker_core_types::dlq::DlqErrorCategory;
 
 #[derive(Clone, Default)]
 struct SharedBuffer(Arc<Mutex<Vec<u8>>>);
